requirements.txt
Flask==3.0.3
rsnchat
python-dotenv